+++ space_packet_parser/exceptions.py
@@ -30,3 +30,17 @@ class CalibrationError(Exception):
 class InvalidParameterTypeError(Exception):
     """Error raised when someone is using an invalid ParameterType element"""
     pass
+
+
+class UnrecognizedPacketTypeError(Exception):
+    """Error raised when we can't figure out which kind of packet we are dealing with based on the header"""
+
+    def __init__(self, *args, partial_data: dict = None):
+        """
+        Parameters
+        ----------
+        partial_data : dict, Optional
+            Data parsed so far (for debugging at higher levels)
+        """
+        super().__init__(*args)
+        self.partial_data = partial_data